+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+##### 1.0.0-beta.1 - 10 May 2016
+
+###### Breaking changes
+- Now depends on js-data 3.0.0-beta.5
+- Now you must import like this:
+
+    ```js
+    // CommonJS
+    var JSDataCloudDatastore = require('js-data-cloud-datastore')
+    var CloudDatastoreAdapter = JSDataCloudDatastore.CloudDatastoreAdapter
+    var adapter = new CloudDatastoreAdapter({...})
+    ```
+
+    ```js
+    // ES2015 modules
+    import {CloudDatastoreAdapter} from 'js-data-cloud-datastore'
+    const adapter = new CloudDatastoreAdapter({...})
+    ```
+
+###### Other
+- Upgraded other dependencies
+
 ##### 1.0.0-alpha.3 - 10 March 2016
 
 ###### Other